modification of the decimal digits helper class
I modified the decimal digits helper class regex pattern
so that it would not be possible to start of the number with
a zero like 012345. Made a new DecimalDigits constructor that
doesn't take any arguments, as that was unneccessary. Fixed
the calls to the helper class apropriately. Also
I made it possible to start of a number with a zero, but it should
be followed by a "." in order to meet the regex pattern.