added a check for input not allowing a "."
I implemented the isValidInput method in the DecimalDigits
class, because my regex expression allows the user to enter
just a "." by itself, therefore this methods checks whether
the user put in a period by itself, and sets the error
on the input edittext. This check is implemented both for the
input in the addExpense class and budgetSetting class alertDialog.