added another check to DecimalDigits isValidInput method
Removed the formatting for the textview in the budgetSettingsActivity
once the user sets the budget. Instead I implemented a check that won't
let the user close the dialog if he leaves a trailing dot in his input.
Also added some javadocs and comments for budgetSettingsAcitivty and
DecimalDigits class for easier code review and orientation.