Added txt files and code to extract info from them
Added two more txt files that hold names of the currencies
and their respective symbols. Also added code that extracts
that info into separate arrays that will later be used
to create an array of Currency objects that will be used
to create a listview in the CurrencyActivity.

Encountered a problem that the symbols don't translate correctly
and they might have to converted to unicode later, if the problem
will still persist on the UI.