Refactor the DB to hold day, month, year values
The previous design of the DB did not provide an easy way
of extracting the needed dates from the DB as it held the
whole date in a String format which would make extracting
the day and month values by themselves very complex.
Therefore I refactored the design to make it hold day,
month and year by themselves for easier access later.

-refactor the classes that work with the DB to hold
day, month and year values by themselves.