Make a view for showing the balance
Even though the progress bar shows how much money the
person has spent today and this month, he might not know
whether he is on track with his balance, because he might
have overdrafted or saved money in the previous days. This
is where the balance comes in. It shows whether the person
overdrafted or saved the money in previous days, so that the
person knows whether he can still spend money or whether he
should not spend money anymore even though the daily budget
says that he still can.

Future versions will implement a function for recounting the
daily budget according to how much the person has overdrafted.