Remove the startActivityForResult and other stuff
Now that the values of the current budget and current
currency used is saved in the start up of the app, there
is no point the startActivity for result in the budget
settings activity. There was also a need to optimize a bunch
of other stuff.

-Remove the startActivityForResult in budgetSetting Activity,
substitute it with the regular startActivity
-Implement a setCurrentCurrencyUsed method in the Currency
class
-Save the values into SharedPreferences right in the Currency
list activity via the newly create setCurrentCurrencyUsed method
-Remove the onActivityResult callback method in BudgetSettings
as the values are being saved in the CurrencyListActivity now
-Impelement the onResume() callback method in budgetSettings
Activity in order to refresh the view with new currencies that
have been set in the Curren List Activity