All except the dailyB include the monthly expenses
On the initial start-up in the middle of the month
the user shouldn't be limited to only setting expenses
after the budget setting date. In case the user does choose
to implement an expense that happened before the date that
the budget was set - the balance changes, and the monthly
budget gets influenced, but not the daily budget.

The user will be offered to reduce the daily budget in case
he overdrafts the balance by clicking a button later

-change logic to include all monthly expenses into the
balance and the daily budget even if the user started using
the app not on the first of the month