Remove the case for asking partial budget
Right now I figured that the code gets too complicated
in case I try to impelement the case where the user
must set the partial budget in case he starts the app
intially in the middle of the month. This feature will
be added in the next version after the initial release.

-Remove the condition on start-up for the start-up in case
of usage in the middle of the month and the user's first
monthly budget will be partial.