Implement the rest of the logic for balance
In previous versions in case the balance was negative and
the button to recalculate the daily budget was pressed
the daily budget got recounted, however the balance was
left negative as it is. This is fixed now and the balance
becomes null in case it is recounted.

-add a timestamp to all expenses
-implement the alternative logic for counting the balance in
getBalance() method in case the budget was recounted this month
already
-add a timestamp to the dailyBudget in case it was recounted