Implement the monthly budget overdraft case
Before - in case the user overdrafted the monthly budget
then the balance would show a warning button - and then
in case you clicked on it - it would seriously mess up a
daily budget. Fixed this by implementing an overdraft
warning button near the motnhly budget in case of monthly
budget overdraft. The button offers the user to raise their
monthly budget.

-implement a check to show the balanceOverdraftButton only
in case the current monthly budget is bigger than all of the
monthly expenses
-implement a warning button to be shown near the monthly budget
in case the user overdrafts the motnhly budget
-offer the user via prompt by clicking the button to raise their
monthly budget to the level higher than all of the monthly
expenses