Make the monthly budget not less than 100
In order to avoid problems with having a monthly budget
that is too small and would be causing problems with the
daily budget there has to be a check for the minimum value
allowed in the motnhly budget. This might be altered in
case the users report the 100 is too small of a value.

-implement a check on the monthly budget setting to be not
less than 100, else set an error on the editText