Add a separate layout for no expenses
In case the user tries to enter ExpandableExpenseActivity
and there are no expenses then the size of the list that
holds all expense will be zero and this will cause problems
later. This can be solved by adding a separate layout and a
check for the size of the list.

-Add a check for the size of the list with expenses
-add a separate layout for an empty expenses list