Implement the delete expense button
In case the user clicks the delete expense button
the chosen expense must be deleted from the database
and the user must return to the list of expenses.

-implement the delete button in the ShowExpense
Activity
-implement a prompt to be asked upon deleting an
expense. In case of positive response - delete the
expense from the database and return to the
ExpandableExpenseActivity and refresh it by setting
a new adapter with a new dataset.
-make the AddExpenseActivity and ShowExpenseActivity
leave no trace in the Activity stack.