make SaveExpenseActivity execute different code
The SaveExpenseActivity must execute different code
paths depending on its parent Activity. In case the
its parent Activit is the mainActivity then SaveExpense
Activity must create a new expense - in case its parent
Activity is ShowExpense - then it must modify it.

- make SaveExpenseActivity execute different code paths
depending on the key passed to it by an intent from the
parent Activity.