Bug fix: back button in the CurrencyListActivity
The user was not able to back out from the CurrencyListActivity
when he called it from the BudgetSettingsActivity. While it is
correct that the user shouldn't have been able to back out of it
from the initial start of the app - they should have the ability
to back out of it otherwise.

-implement an ACTIVITY_COMING_FROM_KEY in the CurrencyListActivity
in order to check from which activity the user comes from and
implement a different behaviour for the back button from there.